# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface GoogleLoginRequest {
  credential: string;
}

export interface User {
  id: number;
  googleId?: string | null;
  fullName: string;
  email: string;
  profileImage?: string | null;
  authProvider: string;
  createdAt: string;
  updatedAt: string;
}

export type Logout200 = {
  message: string;
};