# lib/api-zod/src/generated/types/user.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface User {
  id: number;
  googleId?: string | null;
  fullName: string;
  email: string;
  profileImage?: string | null;
  authProvider: string;
  createdAt: Date;
  updatedAt: Date;
}